+++ src/frontend/src/pages/Index/Settings/AccountSettings/AccountDetailPanel.tsx
@@ -4,12 +4,12 @@ import { Badge, Group, Stack, Table } from '@mantine/core';
 import { IconEdit, IconKey, IconUser } from '@tabler/icons-react';
 import { useMemo } from 'react';
 
+import { ActionButton } from '@lib/components/ActionButton';
+import { YesNoUndefinedButton } from '@lib/components/YesNoButton';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import type { ApiFormFieldSet } from '@lib/types/Forms';
 import { useNavigate } from 'react-router-dom';
 import { useShallow } from 'zustand/react/shallow';
-import { ActionButton } from '../../../../../lib/components/ActionButton';
-import { YesNoUndefinedButton } from '../../../../../lib/components/YesNoButton';
 import { ActionDropdown } from '../../../../components/items/ActionDropdown';
 import { StylishText } from '../../../../components/items/StylishText';
 import { useEditApiFormModal } from '../../../../hooks/UseForm';
